+++ packages/widgetbook/lib/src/addons/addons.dart
@@ -4,4 +4,5 @@ export 'device_frame_addon/addon.dart';
 export 'localization_addon/addon.dart';
 export 'text_scale_addon/addon.dart';
 export 'theme_addon/addon.dart';
+export 'time_dilation_addon/addon.dart';
 export 'use_case_builder_addon/addon.dart';