+++ packages/widgetbook/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **FEAT**: Add Use-case Builder Addon. ([#880](https://github.com/widgetbook/widgetbook/pull/880))
+
 ## 3.2.0
 
  - **BREAKING**: `DeviceFrameAddon.devices` is no longer nullable, in favor of the new `NoneDevice`. ([#854](https://github.com/widgetbook/widgetbook/pull/854))