+++ packages/widgetbook/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - **FEAT**: Add Use-case Builder Addon. ([#880](https://github.com/widgetbook/widgetbook/pull/880))
+- **FEAT**: Add Experimental Time Dilation Addon. ([#887](https://github.com/widgetbook/widgetbook/pull/887))
 - **REFACTOR**: Change `WidgetbookState.knobs` type from `Map<String, Knob>` to `KnobsRegistry`. ([#885](https://github.com/widgetbook/widgetbook/pull/885))
 - **REFACTOR**: Deprecate `WidgetbookState.updateKnobValue` in favor of `WidgetbookState.knobs.updateValue`. ([#885](https://github.com/widgetbook/widgetbook/pull/885))